+++ source/DSCResources/DSC_Computer/DSC_Computer.psm1
@@ -1,8 +1,6 @@
 [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSAvoidGlobalVars', '', Scope = 'Function')]
 [Diagnostics.CodeAnalysis.SuppressMessageAttribute('PSUseDeclaredVarsMoreThanAssignments', 'DSCMachineStatus', Justification = 'GlobalDsc Variable can be ignored')]
-param
-(
-)
+param ()
 
 $modulePath = Join-Path -Path (Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent) -Parent) -ChildPath 'Modules'
 
@@ -269,7 +267,7 @@ function Set-TargetResource
 
                 if ($computerObject)
                 {
-                    Delete-ADSIObject -Path $computerObject.Path -Credential $Credential
+                    Remove-ADSIObject -Path $computerObject.Path -Credential $Credential
                     Write-Verbose -Message ($script:localizedData.DeletedExistingComputerObject -f $Name, $computerObject.Path)
                 }
 
@@ -756,7 +754,7 @@ function Get-ADSIComputer
     .PARAMETER Credential
         Credential to authenticate to the domain.
 #>
-function Delete-ADSIObject
+function Remove-ADSIObject
 {
     [CmdletBinding()]
     param